requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
bcrypt==4.0.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.7
distro==1.9.0
Django==5.2.4
exceptiongroup==1.3.0
fastapi==0.116.1
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
image==1.5.33
jiter==0.10.0
PyJWT==2.8.0
numpy==2.2.6
openai==1.98.0
opencv-python==4.12.0.88
pandas==2.3.1
passlib==1.7.4
pillow==11.3.0
playwright==1.54.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
pyperclip==1.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlparse==0.5.3
starlette==0.47.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.2
python-multipart==0.0.20
